+++ spec/classes/config_spec.rb
@@ -9,6 +9,7 @@ describe 'apics::config' do
       'group'                           => 'oracle',
       'gateway_props_path'              => '/opt/oracle/installer/gateway-props.json',
       'logical_gateway_id'              => 100,
+      'logical_gateway'                 => :undef,
       'management_service_url'          => 'https://test.apiplatform.ocp.example.com',
       'idcs_url'                        => 'https://idcs.example.com/oauth2/v1/token',
       'request_scope'                   => 'https://apiplatform.example.com.apiplatform offline_access',
@@ -44,6 +45,12 @@ describe 'apics::config' do
           'mode'    => '0440',
           'content' => file_fixture('gateway-props.json'),
         )
+      end
+
+      context "with logical_gateway => 'Test Gateway'" do
+        let(:params) { super().merge('logical_gateway' => 'Test Gateway') }
+
+        it { is_expected.to contain_file('/opt/oracle/installer/gateway-props.json').with_content(%r{"logicalGateway": "Test Gateway"}) }
       end
 
       context "with gateway_node_description => 'Test Node'" do